refactor: centralize permissions and filtering
