+++ apps/blocks/blocks/table/table_block.py
@@ -2,7 +2,7 @@ from django.shortcuts import render
 from apps.blocks.models.block import Block
 from apps.blocks.models.block_column_config import BlockColumnConfig
 from apps.blocks.models.block_filter_config import BlockFilterConfig
-from apps.blocks.helpers.permissions import (
+from apps.workflow.permissions import (
     get_readable_fields_state,
     get_editable_fields_state,
 )
