+++ apps/blocks/helpers/filter_config.py
@@ -1,27 +1,10 @@
 from apps.blocks.models.block_filter_config import BlockFilterConfig
-from apps.blocks.registry import block_registry
 
 
 def get_user_filter_config(user, block):
+    """Return the default filter configuration for a user and block."""
     config = BlockFilterConfig.objects.filter(
         block=block, user=user, is_default=True
     ).first()
     return config.values if config else {}
-
-
-def apply_filter_registry(table_name, queryset, filters, user):
-    block = block_registry.get(table_name)
-
-    # Dynamic per-instance filter schema
-    if block and hasattr(block, "get_filter_schema"):
-        schema = block.get_filter_schema(user)
-    else:
-        schema = {}
-
-    print(schema)
-
-    for key, config in schema.items():
-        if key in filters and filters[key] is not None:
-            queryset = config["handler"](queryset, filters[key])
-    return queryset
