--- apps/blocks/helpers/permissions.py
@@ -1,8 +1,0 @@
-from apps.workflow.permissions import get_readable_fields_state, get_editable_fields_state
-
-__all__ = ["get_readable_fields_state", "get_editable_fields_state"]
-
-
-## USAGE ##
-# from apps.blocks.helpers.permissions import get_readable_fields_state
-# fields = get_readable_fields_state(user, model, instance)
