+++ apps/blocks/services/__init__.py
@@ -1,0 +1,1 @@
+"""Service layer for block related operations."""
